feat: better blog seo
